Conflicts:
	app/controllers/manuscript_manager_templates_controller.rb
	app/models/paper.rb
	db/schema.rb

+++ app/assets/javascripts/controllers/manuscript_manager_template_index_controller.js.coffee
@@ -1,8 +1,11 @@
 ETahi.ManuscriptManagerTemplateIndexController = Ember.ArrayController.extend
+  canDeleteTemplates: Ember.computed.gt('model.length', 1)
+
   actions:
     addTemplate: ->
       @transitionToRoute('manuscript_manager_template.new')
 
     destroyTemplate: (template) ->
-      template.destroyRecord().then =>
-        @get('model').removeObject(template)
+      if @get('canDeleteTemplates')
+        template.destroyRecord().then =>
+          @get('model').removeObject(template)
